+++ pgmpy/inference/__init__.py
@@ -1,10 +1,8 @@
 from .base import Inference
-from .ExactInference import VariableElimination
 from .ExactInference import BeliefPropagation
-from .Sampling import BayesianModelSampling, GibbsSampling
+from .ExactInference import VariableElimination
 from .dbn_inference import DBNInference
 from .mplp import Mplp
-from . import continuous
 
 __all__ = ['Inference',
            'VariableElimination',
