+++ pgmpy/sampling/base.py
