+++ pgmpy/tests/test_inference/test_continuous_sampling.py
@@ -2,10 +2,9 @@ import unittest
 
 import numpy as np
 
-from pgmpy.inference.continuous import (HamiltonianMC as HMC, HamiltonianMCda as HMCda, ModifiedEuler,
-                                        GradLogPDFGaussian, NoUTurnSampler as NUTS,
-                                        NoUTurnSamplerDA as NUTSda)
 from pgmpy.factors.continuous import JointGaussianDistribution as JGD
+from pgmpy.sampling import (HamiltonianMC as HMC, HamiltonianMCda as HMCda, GradLogPDFGaussian, NoUTurnSampler as NUTS,
+                            NoUTurnSamplerDA as NUTSda)
 
 
 class TestHMCInference(unittest.TestCase):
